requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.0.0
xlrd>=2.0.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pymongo>=4.0.0          # ← ADD THIS
python-jose[cryptography]>=3.3.0  # ← ADD THIS
passlib[bcrypt]>=1.7.4   # ← ADD THIS
python-dotenv>=1.0.0     # ← ADD THIS (MOST IMPORTANT!)